+++ Changelog.md
@@ -17,7 +17,7 @@ Types of changes:
 ### Added
 
 - Framework to "produce" and "consume" frames from either videos or directories of images
-
+- Renamed package from `dfki_sl_videotools` to `slvideotools`
 
 ## [0.1] - 2022-01-26